requirements.txt
alembic==1.14.0
amqp==5.3.1
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1732732445543/work
anyio==4.8.0
appdirs==1.4.4
archspec @ file:///croot/archspec_1709217642129/work
asn1crypto==1.5.1
async-property==0.2.2
azure-core==1.32.0
azure-identity==1.19.0
azure-storage-blob==12.24.0
azure-storage-file-datalake==12.18.0
backoff==2.2.1
billiard==4.2.1
blinker==1.9.0
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
boto3==1.35.98
botocore==1.35.98
Brotli @ file:///croot/brotli-split_1714483155106/work
cachelib==0.9.0
cachetools==5.5.0
celery==5.4.0
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi @ file:///croot/cffi_1726856441404/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
conda @ file:///croot/conda_1733841818148/work
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1733258728466/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1727775630457/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1731366181659/work
cryptography @ file:///croot/cryptography_1732129644106/work
dash==2.17.0
dash-core-components==2.0.0
dash-embedded==2.18.1
dash-enterprise-libraries==1.6.0
dash-html-components==2.0.0
dash-notes==0.0.3
dash-snapshots==2.2.8
dash-table==5.0.0
dash-user-analytics==0.0.2
dash_ag_grid==31.3.0
dash_canvas==0.1.0
dash_design_kit==1.14.0
dash_enterprise_auth==0.2.4
data_sources==0.2.1
databricks-sdk==0.40.0
databricks-sql-connector==2.9.6
de-client==1.9.0
de_logging==1.0.0
deferred-import==0.1.0
deprecation==2.1.0
diskcache==5.6.3
distro @ file:///croot/distro_1714488253808/work
duckdb==1.1.3
et_xmlfile==2.0.0
fakeredis==1.0.3
filelock==3.16.1
Flask==2.2.5
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==2.5.1
frozendict @ file:///croot/frozendict_1713194832637/work
gevent==24.11.1
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gql==3.5.0
graphql-core==3.2.3
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==20.0.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
idna @ file:///croot/idna_1714398848350/work
imageio==2.37.0
importlib_metadata==8.5.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jwcrypto==1.5.6
kombu==5.4.2
lazy_loader==0.4
libmambapy @ file:///croot/mamba-split_1732896306512/work/libmambapy
lz4==4.3.3
Mako==1.3.8
MarkupSafe==3.0.2
menuinst @ file:///croot/menuinst_1731364904452/work
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
narwhals==1.22.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.1
oauthlib==3.2.2
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.2.3
pg8000==1.31.2
pillow==11.1.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
plotly==5.24.1
pluggy @ file:///croot/pluggy_1733169602837/work
portalocker==2.10.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT==2.10.1
pyOpenSSL==24.3.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
python-keycloak==5.1.1
python-slugify==8.0.4
pytz==2024.2
redis==5.2.1
requests @ file:///croot/requests_1730999120400/work
requests-toolbelt==1.0.0
retrying==1.3.4
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
s3transfer==0.10.4
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
scramp==1.4.5
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.12.4
snowflake-sqlalchemy==1.5.3
sortedcontainers==2.4.0
SQLAlchemy==1.4.54
sqlalchemy-bigquery==1.12.0
tenacity==9.0.0
text-unidecode==1.3
threadpoolctl==3.6.0
thrift==0.16.0
tifffile==2025.3.30
tomlkit==0.13.2
tqdm @ file:///croot/tqdm_1724853939799/work
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typing_extensions==4.12.2
tzcron==1.0.0
tzdata==2024.2
urllib3 @ file:///croot/urllib3_1727769808118/work
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.2.3
wheel==0.44.0
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
zstandard @ file:///croot/zstandard_1731356346222/work
